requirements.txt
certifi==2023.11.17
charset-normalizer==2.1.1
idna==3.6
kafka-python==2.0.2
numpy==1.26.3
pandas==2.2.0
pyarrow==7.0.0
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==0.9.1
tzdata==2023.4
urllib3==1.26.7
apache-airflow==2.6.0
cassandra-driver>=3.29.0
pyspark==3.4.1